1. using glog to print log 2. fix some bug

+++ src/utility/include/constants.h
@@ -10,7 +10,7 @@ constexpr uint8_t varint_16bits = 0xfd;
 constexpr uint8_t varint_32bits = 0xfe;
 constexpr uint8_t varint_64bits = 0xff;
 
-constexpr uint32_t satoshi_per_bitcoin = 100000000;
+constexpr uint64_t satoshi_per_bitcoin = 100000000;
 constexpr uint64_t max_satoshi_amount = 21000000 * satoshi_per_bitcoin;
 
 #endif // BTCLITE_CONSTANTS_H
